Add a handler for checking X509 subject DNs in a request.
This adds a facility for wrapping an http.Handler to support
checking for the correct X509 subject DN in the request before passing it
along. The imagined use case is that you terminate SSL in e.g. Nginx and
insert a header containing the subject DN of the client cert presented.
The app uses that information to do access control on a per-http-handler
basis.

This extracts code from an internal Stripe project.